Mode combat
Menu static afin de palier au probleme de progmem
Mode combat non fonctionelle a debugger

+++ AnnimStartFight.cpp
@@ -13,7 +13,7 @@
   
   IContexte * AnnimStartFight::Update(Gamebuino gb, IContexte *ctx)
   {
-    int currentFrame = gb.frameCount;
+    uint8_t currentFrame = gb.frameCount;
     //l'animation sera de faire passer l'ecran du noir au blanc 
 
     if(this->nbFrameAnnim <1)
@@ -34,9 +34,9 @@
     //#define LCDWIDTH 48
     //#define LCDHEIGHT 84
     
-    for(int i=0;i<15;i++)
+    for(uint8_t i=0;i<15;i++)
     {
-      for(int y=0;y<15;y++)
+      for(uint8_t y=0;y<15;y++)
       {
         if(isWhite)
         {
